Desktop/Django/project1/requirements.txt
Django
django-cors-headers
djangorestframework
django-environ
psycopg2
django-filter
redis